fix: Use AppProvidedSearchEngine when extracting the engine configurations for better accuracy.
This change will make it so that we process the received configuration with AppProvidedSearchEngine and ensure that the values that we display on the UI are the same as the application would have internally.

This also makes it so that we get the URL data at the same time, which simplifies the display mechanism and calls to the back end.